RNG-0: unify gameplay RNG onto Java LegacyRandomSource core (mc::rng)
Replace the scattered 32-bit Numerical-Recipes LCG copies (and the weather
xorshift) with a single deterministic source, src/gameplay/Random.hpp
(namespace mc::rng), transcribed from Java's LegacyRandomSource / java.util.Random:
48-bit state, x25214903917 +11, high-bit next(bits), nextInt() with power-of-two
multiply and rejection sampling otherwise, nextFloat/nextDouble/nextGaussian, and
seedFromValue() for the setSeed scramble.

- Widen every gameplay RNG carrier from 32-bit to the 48-bit state (stored u64):
  SimpleEntity.rngState, PersistentEntity.rngState, the per-system randomState_
  fields (entity loot / session loot+trample / spawner / weather), and the AI/loot
  function signatures (onSpawn/tick/onAttacked/chooseAngerTime/rollLoot and the
  MiningSystem/BlockBehavior drop fn-ptrs).
- Replace ~15 naked `% N` draws with mc::rng::nextInt(state, N), killing the
  low-bit modulo bias (MobBrain's `% 10` retarget no longer locks to state parity),
  and route the existing `>> 8 % N` loot rolls through the same API.
- Bump the save format: worldVersion 19 -> 20, ENTITY block 4 -> 5, region CCNK
  5 -> 6. Versions >= 5/6 persist the u64 rngState; older records read the u32 and
  zero-extend it, a clean migration (the sequence changed algorithm anyway).
- Leave the decorative generators (particles, audio, ambient music, renderer)
  untouched: they are non-authoritative and out of scope.

Tests: rng_golden_test pins mc::rng against java.util.Random (canonical literals
+ an independent transcription over many seeds, including the rejection boundary
and nextInt(10) uniformity). rng_scope_guard_test asserts the decorative sources
never reach for mc::rng. save_repository_test gains a version-4 legacy region
entity carrying a u32 rngState to prove the widening migration round-trips.
level_test's golden pose digest and version_manifest_test's version pin are
recaptured for the new stream/version. ctest 179/179.

+++ src/gameplay/BlockBehavior.hpp
@@ -180,7 +180,7 @@ struct NeighborUpdateContext final {
 };
 
 using GetDropsFn =
-    MinedDrops (*)(world::Block, const ItemStack&, std::uint32_t&, int, bool);
+    MinedDrops (*)(world::Block, const ItemStack&, std::uint64_t&, int, bool);
 using GetShapeFn = world::BlockShape (*)(world::BlockState);
 using GetStateForPlacementFn = std::optional<world::BlockState> (*)(const PlacementBehaviorContext&);
 using UseItemOnFn = void (*)(const InteractionBehaviorContext&);
@@ -337,7 +337,7 @@ template <class Fn>
 // This is behaviour-identical to calling minedDrops directly, which is what the
 // parity harness asserts across every block and tool.
 [[nodiscard]] inline MinedDrops dispatchBlockDrops(core::BlockId id, const ItemStack& tool,
-                                                   std::uint32_t& randomState, int age = 0,
+                                                   std::uint64_t& randomState, int age = 0,
                                                    bool doubledSlab = false) {
     const auto& behavior = behaviorFor(id);
     if (!behavior.prefilter.has(BlockBehaviorBit::HasDrops) || behavior.getDrops == nullptr) {
