+++ src/persistence/SaveRepository.hpp
@@ -78,7 +78,9 @@ struct PersistentEntity final {
     float health = 0.0F;
     std::int32_t angerTicks = 0;
     std::uint32_t ageTicks = 0U;
-    std::uint32_t rngState = 0U;
+    // The 48-bit mc::rng state (stored in a u64). Widened from 32-bit with
+    // RNG-0; entity block / region chunk record bumped a version to persist it.
+    std::uint64_t rngState = 0U;
     // Entity#fireTicks: a creature saved mid-burn reopens still ablaze. Added in
     // entity block version 2; a version-1 record reads it as zero (not on fire).
     std::int32_t fireTicks = 0;
